+++ redisson/src/main/java/org/redisson/RedissonStream.java
@@ -27,6 +27,7 @@ import org.redisson.client.protocol.RedisCommands;
 import org.redisson.client.protocol.StreamEntryStatus;
 import org.redisson.client.protocol.decoder.*;
 import org.redisson.command.CommandAsyncExecutor;
+
 import java.util.*;
 import java.util.concurrent.TimeUnit;
