+++ redisson/src/test/java/org/redisson/rx/RedissonStreamRxTest.java
@@ -1,0 +1,64 @@
+package org.redisson.rx;
+
+import org.junit.jupiter.api.Test;
+import org.redisson.RedisDockerTest;
+import org.redisson.api.RStream;
+import org.redisson.api.RStreamRx;
+import org.redisson.api.stream.*;
+
+import java.util.Map;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+public class RedissonStreamRxTest extends BaseRxTest {
+
+    @Test
+    public void testReadEmptyRx() {
+        RStreamRx<String, String> stream = redisson.getStream("test");
+        Map<StreamMessageId, Map<String, String>> result = stream.read(
+                StreamReadArgs.greaterThan(new StreamMessageId(0, 0)).count(10)).blockingGet();
+        assertThat(result).isEmpty();
+    }
+
+    @Test
+    public void testReadGroupEmptyRx() {
+        RStream<String, String> stream = RedisDockerTest.redisson.getStream("test");
+        stream.createGroup(StreamCreateGroupArgs.name("testGroup").makeStream());
+        stream.add(StreamAddArgs.entry("1", "1"));
+        stream.readGroup("testGroup", "consumer1", StreamReadGroupArgs.neverDelivered());
+
+        RStreamRx<String, String> streamRx = redisson.getStream("test");
+        Map<StreamMessageId, Map<String, String>> result = streamRx.readGroup(
+                "testGroup", "consumer1", StreamReadGroupArgs.neverDelivered()).blockingGet();
+        assertThat(result).isEmpty();
+    }
+
+    @Test
+    public void testReadMultiEmptyRx() {
+        RStreamRx<String, String> stream = redisson.getStream("test1");
+        Map<String, Map<StreamMessageId, Map<String, String>>> result = stream.read(
+                StreamMultiReadArgs.greaterThan(new StreamMessageId(0, 0), "test2", new StreamMessageId(0, 0)).count(10))
+                .blockingGet();
+        assertThat(result).isEmpty();
+    }
+
+    @Test
+    public void testReadGroupMultiEmptyRx() {
+        RStream<String, String> stream1 = RedisDockerTest.redisson.getStream("test1");
+        RStream<String, String> stream2 = RedisDockerTest.redisson.getStream("test2");
+        stream1.createGroup(StreamCreateGroupArgs.name("testGroup").makeStream());
+        stream2.createGroup(StreamCreateGroupArgs.name("testGroup").makeStream());
+        stream1.add(StreamAddArgs.entry("1", "1"));
+        stream2.add(StreamAddArgs.entry("1", "1"));
+        stream1.readGroup("testGroup", "consumer1", StreamMultiReadGroupArgs.greaterThan(
+                StreamMessageId.NEVER_DELIVERED, "test2", StreamMessageId.NEVER_DELIVERED));
+
+        RStreamRx<String, String> streamRx = redisson.getStream("test1");
+        Map<String, Map<StreamMessageId, Map<String, String>>> result = streamRx.readGroup(
+                "testGroup", "consumer1", StreamMultiReadGroupArgs.greaterThan(
+                        StreamMessageId.NEVER_DELIVERED, "test2", StreamMessageId.NEVER_DELIVERED))
+                .blockingGet();
+        assertThat(result).isEmpty();
+    }
+
+}
